Divide a code on auth and deal records
